feat: support select component build target
