+++ packages/rax-template/template/component/_gitignore.ejs
@@ -9,4 +9,5 @@ node_modules/
 build/
 dist/
 lib/
+demo/miniapp/components/
 _miniapp/
