--- packages/rax-template/template/component/app.json.ejs
@@ -1,15 +1,0 @@
-{
-  "pages": [
-    "demo/miniapp/index"
-  ],
-  "window": {
-    "enableWK": "YES",
-    "enableDSL": "YES",
-    "enableJSC": "YES",
-    "defaultTitle": "demo",
-    "backgroundColor": "#F5F5F9",
-    "pullRefresh": false,
-    "allowsBounceVertical": true
-  },
-  "debug": true
-}
