+++ packages/rax-template/template/component/build.json.ejs
@@ -3,7 +3,7 @@
     [
       "rax-plugin-component",
       {
-        "targets": ["web", "weex", "miniapp"]
+        "targets": ["<%- projectTargets.join('", "') %>"]
       }
     ]
   ]
