+++ packages/rax-template/template/component/demo/miniapp/app.js.ejs
