+++ packages/rax-template/template/component/demo/miniapp/app.json.ejs
@@ -1,0 +1,8 @@
+{
+  "pages": [
+    "pages/index"
+  ],
+  "window": {
+    "defaultTitle": "demo",
+  }
+}
