+++ packages/rax-template/template/component/demo/miniapp/pages/index.axml.ejs
