+++ packages/rax-template/template/component/demo/miniapp/pages/index.json.ejs
@@ -1,6 +1,6 @@
 {
   "defaultTitle": "miniapp Rax Text demo",
   "usingComponents": {
-    "my-component": "../../lib/miniapp/index"
+    "my-component": "../component/Target/index"
   }
 }
