--- packages/rax-template/template/component/src/miniapp/index.js.ejs
@@ -1,14 +1,0 @@
-import { createElement, render } from 'rax';
-import View from 'rax-view';
-import Text from 'rax-text';
-
-const MyComponent = props => {
-  return (
-    <View>
-      <Text>Hello World</Text>
-    </View>
-  );
-};
-
-export default MyComponent;
-
