--- packages/rax-template/template/component/src/web/index.jsx.ejs
@@ -1,17 +1,0 @@
-import { createElement } from 'rax';
-import styles from './style';
-
-const MyComponent = props => {
-  const { style, ...rest } = props;
-  let styleProps = {
-    ...styles.initial,
-    ...style
-  };
-  return (
-    <span {...rest} style={styleProps}>
-      Hello World
-    </span>
-  );
-};
-
-export default MyComponent;
