--- packages/rax-template/template/component/src/web/style.js.ejs
@@ -1,12 +1,0 @@
-const styles = {
-  initial: {
-    border: '0 solid black',
-    boxSizing: 'border-box',
-    display: 'flex',
-    flexDirection: 'column',
-    alignContent: 'flex-start',
-    flexShrink: 0
-  }
-};
-
-export default styles;
