--- packages/rax-template/template/component/src/weex/index.jsx.ejs
@@ -1,7 +1,0 @@
-import { createElement } from 'rax';
-
-const MyComponent = props => {
-  return <text {...props}>Hello World</text>;
-};
-
-export default MyComponent;
