--- packages/rax-template/template/component/src/weex/style.js.ejs
@@ -1,3 +1,0 @@
-const styles = {};
-
-export default styles;
