--- packages/rax-template/template/component/tsconfig.json.ejs
@@ -1,23 +1,0 @@
-{
-  "compilerOptions": {
-    "module": "commonjs",
-    "target": "es5",
-    "jsx": "preserve",
-    "jsxFactory": "createElement",
-    "moduleResolution": "node",
-    "baseUrl": ".",
-    "rootDir": "src",
-    "downlevelIteration": true,
-    "alwaysStrict": true,
-    "outDir": "lib",
-    "strict": true,
-    "preserveConstEnums": true,
-    "experimentalDecorators": true,
-    "noImplicitReturns": true,
-    "noUnusedLocals": true,
-    "noImplicitAny": false,
-    "noImplicitThis": false
-  },
-  "include": ["src/**/*"],
-  "exclude": ["**/__tests__"]
-}
